25-07-14 칸반보드 가로스크롤 개선 및 필터링 기능 추가, 프로필부분 개선, 이슈카드 담당자 추가

+++ src/notification/notification.service.ts
@@ -106,8 +106,6 @@ export class NotificationService {
   }
 
   async deleteUserNotification(userId: string, notificationId: string) {
-    await this.userNotificationRepository.delete(
-      {userId, notificationId}
-    )
+    await this.userNotificationRepository.delete({ userId, notificationId });
   }
 }
